Fix resolved mount points for non-overlayfs filesystem (#6723)

+++ pkg/security/tests/testdrive.go
@@ -42,6 +42,10 @@ func (td *testDrive) Path(filename string) (string, unsafe.Pointer, error) {
 }
 
 func newTestDrive(fsType string, mountOpts []string) (*testDrive, error) {
+	return newTestDriveWithMountPoint(fsType, mountOpts, "")
+}
+
+func newTestDriveWithMountPoint(fsType string, mountOpts []string, mountPoint string) (*testDrive, error) {
 	backingFile, err := ioutil.TempFile("", "secagent-testdrive-")
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create testdrive backing file")
@@ -51,10 +55,12 @@ func newTestDrive(fsType string, mountOpts []string) (*testDrive, error) {
 		return nil, errors.Wrap(err, "failed to close testdrive backing file")
 	}
 
-	mountPoint, err := ioutil.TempDir("", "secagent-testdrive-")
-	if err != nil {
-		os.Remove(backingFile.Name())
-		return nil, errors.Wrap(err, "failed to create testdrive mount point")
+	if len(mountPoint) == 0 {
+		mountPoint, err = ioutil.TempDir("", "secagent-testdrive-")
+		if err != nil {
+			os.Remove(backingFile.Name())
+			return nil, errors.Wrap(err, "failed to create testdrive mount point")
+		}
 	}
 
 	if err := os.Truncate(backingFile.Name(), 1*1024*1024); err != nil {
